requirements.txt
Flask
Flask-SocketIO
rich
cryptography
python-socketio[client]
gunicorn
